+++ fe-util/bulid/webpack.base.conf.js
@@ -1,3 +1,9 @@
+/*
+ * @Author: xiaoguang_10@qq.com
+ * @LastEditors: xiaoguang_10@qq.com
+ * @Date: 2020-06-13 16:05:43
+ * @LastEditTime: 2020-12-14 16:23:52
+ */
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
@@ -9,11 +15,12 @@ function resolve(dir) {
 module.exports = {
   entry: resolve('src/index.js'),
   context: process.cwd(), // 上下文目录
-  mode: 'production',
   output: {
     path: resolve('dist'),
     filename: 'index.min.js',
-    publicPath: '/'
+    publicPath: '/',
+    chunkLoading: false,
+    wasmLoading: false,
   },
 
   plugins: [